+++ src/components/feature/CartSection/utils/getSelectedCartItems.ts
@@ -1,0 +1,10 @@
+import { CartProduct } from "../../../../type/cart";
+
+export const getSelectedCartItems = (
+  cartItems: CartProduct[],
+  selectedCartIds: number[]
+) => {
+  return cartItems.filter((item: CartProduct) =>
+    selectedCartIds.includes(item.id)
+  );
+};
