+++ src/components/feature/Coupon/CouponList/utils/getSelectedCoupons.ts
@@ -1,0 +1,6 @@
+import { CouponResponse } from "../../../../../type/coupon";
+
+export const getSelectedCoupons = (
+  coupons: CouponResponse[],
+  selectedIds: number[]
+) => coupons?.filter((coupon) => selectedIds.includes(coupon.id));
